requirements.txt
Django>=3.1.5,<3.2.0
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.7.5,<2.8.0
django-model-utils>=4.1.1,<4.2.0
django-localflavor>=3.0.1,<3.1.0
python-dateutil>=2.8.1,<2.9.0
requests>=2.25.1,<2.26.0
celery>=5.0.5,<5.1.0
redis>=3.5.3,<3.6.0
django-celery-beat>=2.2.0,<2.3.0

flake8>=3.8.4,<3.9.0
